+++ src/components/index.js
@@ -11,4 +11,4 @@ export { Manage } from './manage'
 export { Stats } from './stats'
 export { LogoPrimary } from './LogoPrimary'
 export { LogoWhite } from './LogoWhite'
-export { Crowdsales } from './crowdsales'
+export { Crowdsales } from './Crowdsales'
